requirements.txt
customtkinter
sounddevice
numpy
requests
Pillow
scipy
